gerente

+++ app/Http/Controllers/PaqueteController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Paquete;
 use App\Models\Servicio;
+use App\Models\Evento;
 
 class PaqueteController extends Controller
 {
@@ -41,8 +42,20 @@ class PaqueteController extends Controller
 
     public function edit(string $id)
     {
+        $aux = False;
+        $eventos = Evento::all();
         $paquete = Paquete::find($id);
-        return view('Paquetes.edit', compact('paquete'));
+        foreach ($eventos as $evento) {
+            if  ($evento->paquete_id == $paquete->id && $evento->estatus == "SinConfirmar") {
+                $aux = True;
+                break;
+            }
+        }
+        if ($aux) {
+            return redirect()->back()->with('alert', 'Oh no! No es posible editar el paquete.');
+        } else {
+            return view('Paquetes.edit', compact('paquete'));
+        }
     }
 
     public function update(Request $request, string $id)
@@ -59,8 +72,20 @@ class PaqueteController extends Controller
 
     public function destroy(string $id)
     {
+        $aux = False;
+        $eventos = Evento::all();
         $paquete = Paquete::find($id);
-        $paquete->delete();
-        return redirect(route('sistema.gerente'));
+        foreach ($eventos as $evento) {
+            if  ($evento->paquete_id == $paquete->id) {
+                $aux = True;
+                break;
+            }
+        }
+        if ($aux) {
+            return redirect()->back()->with('alert', 'Oh no! No es posible eliminar el paquete ya que se encunetra en un evento activo.');
+        } else {
+            $paquete->delete();
+            return redirect(route('sistema.gerente'));
+        }
     }
 }
