+++ app/Http/Controllers/ServicioController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Servicio;
+use App\Models\Evento;
 use Illuminate\Http\Request;
 
 class ServicioController extends Controller
@@ -35,8 +36,27 @@ class ServicioController extends Controller
 
     public function edit(string $id)
     {
+        $aux = False;
+        $eventos = Evento::all();
         $servicio = Servicio::find($id);
-        return view('Servicios.edit', compact('servicio'));
+
+        foreach ($eventos as $evento) {
+            if ($evento->estatus == "SinConfirmar") {
+                $servicios_eventos = $evento->servicios;
+                foreach ($servicios_eventos as $serv) {
+                    if ($serv->pivot->servicio_id == $servicio->id) {
+                        $aux = true;
+                        break;
+                    }
+                }
+            }
+        }
+
+        if ($aux) {
+            return redirect()->back()->with('alert', 'Oh no! No es posible editar el servicio.');
+        } else {
+            return view('Servicios.edit', compact('servicio'));
+        }
     }
 
     public function update(Request $request, string $id)
@@ -51,8 +71,25 @@ class ServicioController extends Controller
 
     public function destroy(string $id)
     {
+        $aux = False;
+        $eventos = Evento::all();
         $servicio = Servicio::find($id);
-        $servicio->delete();
-        return redirect(route('sistema.gerente'));
+
+        foreach ($eventos as $evento) {
+            $servicios_eventos = $evento->servicios;
+            foreach ($servicios_eventos as $serv) {
+                if ($serv->pivot->servicio_id == $servicio->id) {
+                    $aux = true;
+                    break;
+                }
+            }
+        }
+
+        if ($aux) {
+            return redirect()->back()->with('alert', 'Oh no! No es posible eliminar el servicio.');
+        } else {
+            $servicio->delete();
+            return redirect(route('sistema.gerente'));
+        }
     }
 }
